# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zorvyn Finance Dashboard API Assignment
  version: 1.0.0
  description: Financial analytics API with JWT auth, RBAC, idempotent
    transactions, soft deletes, and Redis cache-aside acceleration.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Validation failed
        details:
          nullable: true
          description: Optional additional error details
          oneOf:
            - type: string
            - type: object
            - type: array
              items:
                type: object
    Role:
      type: string
      enum:
        - admin
        - analyst
        - viewer
    UserStatus:
      type: string
      enum:
        - active
        - inactive
    TransactionType:
      type: string
      enum:
        - income
        - expense
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
          maxLength: 255
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/UserStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Transaction:
      type: object
      required:
        - id
        - user_id
        - amount
        - type
        - category
        - timestamp
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: string
          example: "2500.00"
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
          maxLength: 100
        note:
          type: string
          nullable: true
          maxLength: 1000
        timestamp:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 42
        total_pages:
          type: integer
          minimum: 1
          example: 3
...
    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: "#/components/schemas/Role"
...
    CreateTransactionRequest:
      type: object
      required:
        - amount
        - type
        - category
      description: Provide idempotency key either in body.idempotency_key or
        Idempotency-Key header.
      properties:
        user_id:
          type: string
          format: uuid
        amount:
          type: string
          pattern: ^(?:0|[1-9]\\d*)(?:\\.\\d{1,2})?$
          example: "2500.00"
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
          minLength: 1
          maxLength: 100
        note:
          type: string
          maxLength: 1000
        timestamp:
          type: string
          format: date-time
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 255
...
  responses:
    BadRequest:
      description: Invalid request payload or query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not allowed for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current state (duplicates, idempotency
        conflict, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Create a user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags:
        - Users
      summary: List users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated users list
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - pagination
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/transactions:
    post:
      tags:
        - Transactions
      summary: Create transaction (idempotent)
      description: If request idempotency matches an existing transaction, returns
        status=existing and HTTP 200.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 255
          description: Optional idempotency key header. Used if body.idempotency_key is
            not provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "200":
          description: Idempotent replay, existing transaction returned
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - status
                      - transaction
                    properties:
                      status:
                        type: string
                        enum:
                          - created
                          - existing
                      transaction:
                        $ref: "#/components/schemas/Transaction"
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - status
                      - transaction
                    properties:
                      status:
                        type: string
                        enum:
                          - created
                          - existing
                      transaction:
                        $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags:
        - Transactions
      summary: List transactions
      description: Viewers are scoped to their own user_id regardless of provided
        query filter.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/TransactionType"
        - in: query
          name: category
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated transaction list
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - pagination
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...